refactor/client_handler: move coin operations into a separate module
