+++ src/client_handler.rs
@@ -9,6 +9,7 @@
 mod auth;
 mod auth_keys;
 mod balance;
+mod coin_operations;
 mod elder_data;
 mod login_packets;
 mod messaging;
@@ -16,7 +17,8 @@ mod messaging;
 use self::{
     auth::{Auth, ClientInfo},
     auth_keys::AuthKeysDb,
-    balance::{Balance, BalancesDb},
+    balance::BalancesDb,
+    coin_operations::CoinOperations,
     elder_data::ElderData,
     login_packets::LoginPackets,
     messaging::Messaging,
@@ -31,11 +33,11 @@ use crate::{
     Config, Result,
 };
 use bytes::Bytes;
-use log::{error, info, trace};
+use log::{error, trace};
 use rand::{CryptoRng, Rng};
 use safe_nd::{
-    Coins, Error as NdError, MessageId, NodePublicId, PublicId, PublicKey, Request, Response,
-    Signature, Transaction, TransactionId, XorName,
+    Coins, Error as NdError, MessageId, NodePublicId, PublicId, Request, Response, Signature,
+    XorName,
 };
 use std::{
     cell::{Cell, RefCell},
@@ -50,10 +52,10 @@ pub const COST_OF_PUT: Coins = Coins::from_nano(1);
 pub(crate) struct ClientHandler {
     id: NodePublicId,
     messaging: Messaging,
+    coin_operations: CoinOperations,
     auth: Auth,
     login_packets: LoginPackets,
     data: ElderData,
-    balances: BalancesDb,
 }
 
 impl ClientHandler {
@@ -78,16 +80,17 @@ impl ClientHandler {
         let messaging = Messaging::new(id.clone(), routing_node);
 
         let auth = Auth::new(id.clone(), auth_keys_db);
+        let coin_operations = CoinOperations::new(id.clone(), balances);
         let login_packets = LoginPackets::new(id.clone(), login_packets_db);
         let data = ElderData::new(id.clone());
 
         let client_handler = Self {
             id,
             messaging,
+            coin_operations,
             auth,
             login_packets,
             data,
-            balances,
         };
 
         Ok(client_handler)
@@ -119,13 +122,17 @@ impl ClientHandler {
                 refund,
             } => {
                 if let Some(refund_amount) = refund {
-                    if let Err(error) = self.deposit(requester.name(), refund_amount) {
+                    if let Err(error) = self
+                        .coin_operations
+                        .deposit(requester.name(), refund_amount)
+                    {
                         error!(
                             "{}: Failed to refund {} coins for {:?}: {:?}",
                             self, refund_amount, requester, error,
                         )
                     };
                 }
+
                 self.messaging
                     .relay_reponse_to_client(src, &requester, response, message_id)
             }
@@ -143,13 +150,15 @@ impl ClientHandler {
                 cost,
             } => {
                 let owner = utils::owner(&client_public_id)?;
-                self.pay(
+                if let Some(action) = self.coin_operations.pay(
                     &client_public_id,
                     owner.public_key(),
                     &request,
                     message_id,
                     cost,
-                )?;
+                ) {
+                    return Some(action);
+                }
 
                 Some(Action::ForwardClientRequest(Rpc::Request {
                     requester: client_public_id,
@@ -174,13 +183,15 @@ impl ClientHandler {
             } => {
                 let owner = utils::owner(&client_public_id)?;
 
-                self.pay(
+                if let Some(action) = self.coin_operations.pay(
                     &client_public_id,
                     owner.public_key(),
                     &request,
                     message_id,
                     cost,
-                )?;
+                ) {
+                    return Some(action);
+                }
 
                 Some(Action::ProxyClientRequest(Rpc::Request {
                     requester: client_public_id,
@@ -330,7 +341,7 @@ impl ClientHandler {
                 destination,
                 amount,
                 transaction_id,
-            } => self.handle_transfer_coins_client_req(
+            } => self.coin_operations.process_transfer_coins_client_req(
                 &client.public_id,
                 destination,
                 amount,
@@ -339,6 +350,7 @@ impl ClientHandler {
             ),
             GetBalance => {
                 let balance = self
+                    .coin_operations
                     .balance(client.public_id.name())
                     .ok_or(NdError::NoSuchBalance);
                 let response = Response::GetBalance(balance);
@@ -349,7 +361,7 @@ impl ClientHandler {
                 new_balance_owner,
                 amount,
                 transaction_id,
-            } => self.handle_create_balance_client_req(
+            } => self.coin_operations.process_create_balance_client_req(
                 &client.public_id,
                 new_balance_owner,
                 amount,
@@ -435,7 +447,10 @@ impl ClientHandler {
             } => {
                 if &src == requester.name() {
                     // Step two - create balance and forward login_packet.
-                    if let Err(error) = self.create_balance(&requester, new_owner, amount) {
+                    if let Err(error) = self
+                        .coin_operations
+                        .create_balance(&requester, new_owner, amount)
+                    {
                         // Refund amount (Including the cost of creating the balance)
                         let refund = Some(amount.checked_add(COST_OF_PUT)?);
 
@@ -474,24 +489,55 @@ impl ClientHandler {
                 new_balance_owner,
                 amount,
                 transaction_id,
-            } => self.handle_create_balance_vault_req(
-                requester,
-                new_balance_owner,
-                amount,
-                transaction_id,
-                message_id,
-            ),
+            } => {
+                let action = self.coin_operations.finalize_create_balance_req(
+                    requester,
+                    new_balance_owner,
+                    amount,
+                    transaction_id,
+                    message_id,
+                );
+                let destination = XorName::from(new_balance_owner);
+                if let Some(Action::RespondToClientHandlers {
+                    rpc:
+                        Rpc::Response {
+                            response: Response::Transaction(Ok(transaction)),
+                            ..
+                        },
+                    ..
+                }) = &action
+                {
+                    self.messaging
+                        .notify_destination_owners(&destination, *transaction);
+                }
+                action
+            }
             TransferCoins {
                 destination,
                 amount,
                 transaction_id,
-            } => self.handle_transfer_coins_vault_req(
-                requester,
-                destination,
-                amount,
-                transaction_id,
-                message_id,
-            ),
+            } => {
+                let action = self.coin_operations.finalize_transfer_coins_req(
+                    requester,
+                    destination,
+                    amount,
+                    transaction_id,
+                    message_id,
+                );
+                if let Some(Action::RespondToClientHandlers {
+                    rpc:
+                        Rpc::Response {
+                            response: Response::Transaction(Ok(transaction)),
+                            ..
+                        },
+                    ..
+                }) = &action
+                {
+                    self.messaging
+                        .notify_destination_owners(&destination, *transaction);
+                }
+                action
+            }
             UpdateLoginPacket(updated_login_packet) => self
                 .login_packets
                 .finalize_login_packet_update(requester, &updated_login_packet, message_id),
@@ -552,224 +598,6 @@ impl ClientHandler {
                 );
                 None
             }
-        }
-    }
-
-    fn handle_create_balance_client_req(
-        &mut self,
-        requester: &PublicId,
-        owner_key: PublicKey,
-        amount: Coins,
-        transaction_id: TransactionId,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let request = Request::CreateBalance {
-            new_balance_owner: owner_key,
-            amount,
-            transaction_id,
-        };
-        // For phases 1 & 2 we allow owners to create their own balance freely.
-        let own_request = utils::own_key(requester)
-            .map(|key| key == &owner_key)
-            .unwrap_or(false);
-        if own_request {
-            return Some(Action::VoteFor(ConsensusAction::Forward {
-                request,
-                client_public_id: requester.clone(),
-                message_id,
-            }));
-        }
-
-        let total_amount = amount.checked_add(COST_OF_PUT)?;
-        // When ClientA(owner/app with permissions) creates a balance for ClientB
-        Some(Action::VoteFor(ConsensusAction::PayAndForward {
-            request,
-            client_public_id: requester.clone(),
-            message_id,
-            cost: total_amount,
-        }))
-    }
-
-    fn handle_create_balance_vault_req(
-        &mut self,
-        requester: PublicId,
-        owner_key: PublicKey,
-        amount: Coins,
-        transaction_id: TransactionId,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let (result, refund) = match self.create_balance(&requester, owner_key, amount) {
-            Ok(()) => {
-                let destination = XorName::from(owner_key);
-                let transaction = Transaction {
-                    id: transaction_id,
-                    amount,
-                };
-                self.messaging
-                    .notify_destination_owners(&destination, transaction);
-                (Ok(transaction), None)
-            }
-            Err(error) => {
-                // Refund amount (Including the cost of creating a balance)
-                let amount = amount.checked_add(COST_OF_PUT)?;
-                (Err(error), Some(amount))
-            }
-        };
-
-        Some(Action::RespondToClientHandlers {
-            sender: *self.id.name(),
-            rpc: Rpc::Response {
-                response: Response::Transaction(result),
-                requester,
-                message_id,
-                refund,
-            },
-        })
-    }
-
-    fn handle_transfer_coins_client_req(
-        &mut self,
-        requester: &PublicId,
-        destination: XorName,
-        amount: Coins,
-        transaction_id: TransactionId,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        Some(Action::VoteFor(ConsensusAction::PayAndForward {
-            request: Request::TransferCoins {
-                destination,
-                amount,
-                transaction_id,
-            },
-            client_public_id: requester.clone(),
-            message_id,
-            cost: amount,
-        }))
-    }
-
-    fn handle_transfer_coins_vault_req(
-        &mut self,
-        requester: PublicId,
-        destination: XorName,
-        amount: Coins,
-        transaction_id: TransactionId,
-        message_id: MessageId,
-    ) -> Option<Action> {
-        let (result, refund) = match self.deposit(&destination, amount) {
-            Ok(()) => {
-                let transaction = Transaction {
-                    id: transaction_id,
-                    amount,
-                };
-
-                self.messaging
-                    .notify_destination_owners(&destination, transaction);
-                (Ok(transaction), None)
-            }
-            Err(error) => (Err(error), Some(amount)),
-        };
-
-        Some(Action::RespondToClientHandlers {
-            sender: *self.id.name(),
-            rpc: Rpc::Response {
-                response: Response::Transaction(result),
-                requester,
-                message_id,
-                refund,
-            },
-        })
-    }
-
-    fn withdraw<K: balance::Key>(&mut self, key: &K, amount: Coins) -> Result<(), NdError> {
-        if amount.as_nano() == 0 {
-            return Err(NdError::InvalidOperation);
-        }
-        let (public_key, mut balance) = self
-            .balances
-            .get_key_value(key)
-            .ok_or(NdError::NoSuchBalance)?;
-        balance.coins = balance
-            .coins
-            .checked_sub(amount)
-            .ok_or(NdError::InsufficientBalance)?;
-        self.put_balance(&public_key, &balance)
-    }
-
-    fn deposit<K: balance::Key>(&mut self, key: &K, amount: Coins) -> Result<(), NdError> {
-        let (public_key, mut balance) = self
-            .balances
-            .get_key_value(key)
-            .ok_or_else(|| NdError::NoSuchBalance)?;
-        balance.coins = balance
-            .coins
-            .checked_add(amount)
-            .ok_or(NdError::ExcessiveValue)?;
-
-        self.put_balance(&public_key, &balance)
-    }
-
-    fn put_balance(&mut self, public_key: &PublicKey, balance: &Balance) -> Result<(), NdError> {
-        trace!(
-            "{}: Setting balance to {} for {}",
-            self,
-            balance,
-            public_key
-        );
-        self.balances.put(public_key, balance).map_err(|error| {
-            error!(
-                "{}: Failed to update balance of {}: {}",
-                self, public_key, error
-            );
-
-            NdError::from("Failed to update balance")
-        })
-    }
-
-    // Pays cost of a request.
-    fn pay(
-        &mut self,
-        requester_id: &PublicId,
-        requester_key: &PublicKey,
-        request: &Request,
-        message_id: MessageId,
-        cost: Coins,
-    ) -> Option<()> {
-        trace!("{}: {} is paying {} coins", self, requester_id, cost);
-        match self.withdraw(requester_key, cost) {
-            Ok(()) => Some(()),
-            Err(error) => {
-                trace!("{}: Unable to withdraw {} coins: {}", self, cost, error);
-                self.messaging
-                    .respond_to_client(message_id, request.error_response(error));
-                None
-            }
-        }
-    }
-
-    fn balance<K: balance::Key>(&self, key: &K) -> Option<Coins> {
-        self.balances.get(key).map(|balance| balance.coins)
-    }
-
-    fn create_balance(
-        &mut self,
-        requester: &PublicId,
-        owner_key: PublicKey,
-        amount: Coins,
-    ) -> Result<(), NdError> {
-        let own_request = utils::own_key(requester)
-            .map(|key| key == &owner_key)
-            .unwrap_or(false);
-        if !own_request && self.balances.exists(&owner_key) {
-            info!(
-                "{}: Failed to create balance for {:?}: already exists.",
-                self, owner_key
-            );
-
-            Err(NdError::BalanceExists)
-        } else {
-            let balance = Balance { coins: amount };
-            self.put_balance(&owner_key, &balance)?;
-            Ok(())
         }
     }
 }
